+++ pc-erp/src/main/java/net/caiban/pc/erp/domain/product/ProductDefine.java
@@ -21,15 +21,17 @@ public class ProductDefine extends BaseDomain {
 	 * 
 	 */
 	private static final long serialVersionUID = 1L;
-	private Integer productId;
+	private Long productId;
 	private String details;
-	
-	public Integer getProductId() {
+
+	public Long getProductId() {
 		return productId;
 	}
-	public void setProductId(Integer productId) {
+
+	public void setProductId(Long productId) {
 		this.productId = productId;
 	}
+
 	public String getDetails() {
 		return details;
 	}
