+++ pc-erp/src/main/java/net/caiban/pc/erp/service/events/EventsService.java
@@ -15,12 +15,12 @@ public interface EventsService {
 
 	public Integer[] changeToIntArray(String origin);
 	
-	public Map<String, Integer> filterAppendJoiner(String origin,
+	public Map<String, Long> filterAppendJoiner(String origin,
 			String originId, String append);
 	
 	public void initGmt(Events events, String gmtStartStr, String gmtEndStr);
 	
-	public Integer saveEvent(Events event);
+	public Long saveEvent(Events event);
 	
 	public void appendJoiner(Integer eventId, String inviteAccountId);
 }
