+++ pc-erp/src/main/java/net/caiban/pc/erp/service/sys/SysCompanyService.java
@@ -15,5 +15,5 @@ public interface SysCompanyService {
 
 	public List<SysCompany> query();
 	
-	public SysCompany queryOne(Integer cid);
+	public SysCompany queryOne(Long cid);
 }
