+++ pc-erp/src/main/java/net/caiban/pc/erp/service/sys/impl/SysAppServiceImpl.java
@@ -28,7 +28,7 @@ public class SysAppServiceImpl implements SysAppService {
 	private SysAppMapper sysAppMapper;
 	
 	@Override
-	public SysApp queryByDomain(Integer cid, String domain) {
+	public SysApp queryByDomain(Long cid, String domain) {
 		if(cid==null){
 			return null;
 		}
